refactor(work_engine): modularise cli.py into 7 focused modules (P2.3)
Split the 634-line cli.py monolith into a thin 195-line orchestrator plus
six focused modules under templates/scripts/work_engine/:

  cli_args.py        — argument parsing + state-file constants
  state_io.py        — JSON I/O, WorkState projection, migration hints
  input_builders.py  — build state from tickets, prompts, diffs, files
  hook_bootstrap.py  — assemble the lifecycle hook registry
  emitters.py        — stdout/stderr formatting
  errors.py          — private _CLIError exception
  cli.py             — orchestrator; re-exports public API + monkeypatch
                       surface (e.g. _build_hook_registry) for test compat

Behavior-preserving:
  - 706 work_engine tests + 1708 full-suite tests green
  - No CLI flag, exit code, or stdout/stderr change
  - All existing test monkeypatches still attach to the cli module

Closes P2.3 of road-to-post-pr29-optimize.md.
Roadmap + dashboard updated in the same commit per roadmap-progress-sync.
