+++ agents/roadmaps-progress.md
@@ -6,10 +6,10 @@
 
 ## Overall
 
-**15 / 124 steps done · 12%**
+**16 / 124 steps done · 13%**
 
 ```text
-█████░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░   12%
+█████░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░   13%
 ```
 
 ## Open roadmaps
@@ -20,7 +20,7 @@
 | 2 | [road-to-distribution-and-adoption.md](roadmaps/road-to-distribution-and-adoption.md) | 4 | 13 | 13 | 0 | 0 | 0 | ░░░░░░░░░░ 0% |
 | 3 | [road-to-governance-cleanup.md](roadmaps/road-to-governance-cleanup.md) | 1 | 24 | 24 | 0 | 0 | 0 | ░░░░░░░░░░ 0% |
 | 4 | [road-to-mcp-server.md](roadmaps/road-to-mcp-server.md) | 6 | 29 | 29 | 0 | 0 | 0 | ░░░░░░░░░░ 0% |
-| 5 | [road-to-post-pr29-optimize.md](roadmaps/road-to-post-pr29-optimize.md) | 3 | 24 | 9 | 15 | 0 | 0 | ██████░░░░ 62% |
+| 5 | [road-to-post-pr29-optimize.md](roadmaps/road-to-post-pr29-optimize.md) | 3 | 24 | 8 | 16 | 0 | 0 | ███████░░░ 67% |
 
 ---
 
@@ -69,11 +69,11 @@
 
 ### [road-to-post-pr29-optimize.md](roadmaps/road-to-post-pr29-optimize.md)
 
-**Road to Post-PR-29 Optimize** — 15 / 24 done (62%)
+**Road to Post-PR-29 Optimize** — 16 / 24 done (67%)
 
 | # | Phase | State | Open | Done | Deferred | Cancelled | % |
 |---|---|---|---:|---:|---:|---:|---:|
 | 1 | Housekeeping 1.15.0 | ✅ done | 0 | 13 | 0 | 0 | 100% |
-| 2 | Post-1.15.0 work | 🟡 in progress | 9 | 2 | 0 | 0 | 18% |
+| 2 | Post-1.15.0 work | 🟡 in progress | 8 | 3 | 0 | 0 | 27% |
 | 2 | ICE table | ⬜ empty | 0 | 0 | 0 | 0 | 0% |
